+++ wsd/COOLWSD.cpp
@@ -718,6 +718,7 @@ std::string COOLWSD::ConfigFile = COOLWSD_CONFIGDIR "/coolwsd.xml";
 std::string COOLWSD::ConfigDir = COOLWSD_CONFIGDIR "/conf.d";
 bool COOLWSD::EnableTraceEventLogging = false;
 bool COOLWSD::EnableAccessibility = false;
+bool COOLWSD::EnableMountNamespaces= false;
 FILE *COOLWSD::TraceEventFile = NULL;
 std::string COOLWSD::LogLevel = "trace";
 std::string COOLWSD::LogLevelStartup = "trace";
@@ -1997,6 +1998,7 @@ void COOLWSD::innerInitialize(Application& self)
         { "logging.disable_server_audit", "false" },
         { "browser_logging", "false" },
         { "mount_jail_tree", "true" },
+        { "mount_namespaces", "false" },
         { "net.connection_timeout_secs", "30" },
         { "net.listen", "any" },
         { "net.proto", "all" },
@@ -2166,6 +2168,12 @@ void COOLWSD::innerInitialize(Application& self)
 
     // Experimental features.
     EnableExperimental = getConfigValue<bool>(conf, "experimental_features", false);
+
+    const bool UseMountNamespaces = EnableExperimental && getConfigValue<bool>(conf, "mount_namespaces", false);
+    // attempt this early before starting logging which spawns a thread
+    EnableMountNamespaces = UseMountNamespaces && JailUtil::becomeMountingUser(geteuid(), getegid());
+    if (EnableMountNamespaces)
+        JailUtil::enableMountNamespaces();
 
     EnableAccessibility = getConfigValue<bool>(conf, "accessibility.enable", false);
 
@@ -2531,6 +2539,9 @@ void COOLWSD::innerInitialize(Application& self)
     // For some reason I can't get at this setting in ChildSession::loKitCallback().
     std::string fontsMissingHandling = config::getString("fonts_missing.handling", "log");
     setenv("FONTS_MISSING_HANDLING", fontsMissingHandling.c_str(), 1);
+
+    if (UseMountNamespaces && !EnableMountNamespaces)
+        LOG_WRN("Using mount namespaces failed.");
 
     IsBindMountingEnabled = getConfigValue<bool>(conf, "mount_jail_tree", true);
 #if CODE_COVERAGE
@@ -3477,7 +3488,16 @@ bool COOLWSD::createForKit()
 #elif VALGRIND_COOLFORKIT
     std::string forKitPath = "/usr/bin/valgrind";
 #else
-    std::string forKitPath = parentPath + "coolforkit";
+    std::string forKitPath = parentPath;
+    if (EnableMountNamespaces)
+    {
+        forKitPath += "coolforkitns";
+        args.push_back("--namespace");
+    }
+    else
+    {
+        forKitPath += "coolforkit";
+    }
 #endif
 
     // Always reap first, in case we haven't done so yet.
